+++ __tests__/join/prefix-components-with-info-prop/snapshot.js
@@ -1,15 +1,37 @@
 // Jest Snapshot v1, https://goo.gl/fbAQLP
 
-exports[`E2E join without options test: two-files-with-no-errors 1`] = `
+exports[`E2E join with options test with option: { name: 'prefix-components-with-info-prop', value: 'title' } 1`] = `
 
 openapi: 3.0.0
 info:
   version: 1.0.0
-  title: Example OpenAPI 3 definition.
+  title: Example OpenAPI 3 definition foo.
   description: Information about API
   license:
     name: MIT
     url: https://opensource.org/licenses/MIT
+tags:
+  - name: Foo
+    description: Foo Discription
+    x-displayName: Foo
+  - name: foo_other
+    x-displayName: other
+  - name: Bar
+    description: Bar Discription
+    x-displayName: Bar
+  - name: bar_other
+    x-displayName: other
+x-tagGroups:
+  - name: foo
+    tags:
+      - Foo
+      - foo_other
+    description: Information about API
+  - name: bar
+    tags:
+      - Bar
+      - bar_other
+    description: Information about API
 servers:
   - url: https://redocly.com/v1
 paths:
@@ -39,19 +61,14 @@ paths:
           description: example description
       tags:
         - bar_other
-tags:
-  - name: foo_other
-    x-displayName: other
-  - name: bar_other
-    x-displayName: other
-x-tagGroups:
-  - name: foo
-    tags:
-      - foo_other
-  - name: bar
-    tags:
-      - bar_other
-components: {}
+components:
+  schemas:
+    Example OpenAPI 3 definition foo._some-property:
+      description: foo description
+      type: string
+    Example OpenAPI 3 definition._some-property:
+      description: bar description
+      type: string
 
 openapi.yaml: join processed in <test>ms