+++ packages/cli/src/__mocks__/utils.ts
@@ -9,3 +9,4 @@ export const printUnusedWarnings = jest.fn();
 export const printLintTotals = jest.fn();
 export const getOutputFileName = jest.fn(() => ({ outputFile: 'test.yaml', ext: 'yaml' }));
 export const handleError = jest.fn();
+export const exitWithError = jest.fn();